# addons/Photo-Vault/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Booster Vault API (Non-AI MVP)
  version: 1.0.0
  description: >
    Booster Vault – Memory/Photo Vault (Non-AI MVP). Encrypted media blobs; server stores metadata + signed URLs only.
servers:
  - url: https://api.boostervault.com
    description: Production
  - url: http://localhost:4000
    description: Local
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error code.
          example: VALIDATION_ERROR
          examples: [VALIDATION_ERROR, RECOVERY_REQUIRED]
        message:
          type: string
          example: Invalid request
        statusCode:
          type: number
          nullable: true
          description: Optional HTTP status code field (when returned by framework default errors).
        error:
          type: string
          nullable: true
          description: Optional short error label (when returned by framework default errors).
        details:
          type: object
          additionalProperties: true
          nullable: true
        requestId:
          type: string
          nullable: true
          description: Request correlation id (from x-request-id or generated).
...
    Album:
      type: object
      required: [id, name, createdAt, updatedAt, isDeleted]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        coverMediaId: { type: string, format: uuid, nullable: true }
        isDeleted: { type: boolean }
        deletedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CreateShareDto:
      type: object
      properties:
        expiresInDays:
          type: integer
          minimum: 1
          maximum: 365
          default: 7
    CreateShareRequest:
      type: object
      required: [createShareDto, encryptedAlbumKey, encryptedMediaKeys, iv, kdfParams]
      properties:
        createShareDto: { $ref: "#/components/schemas/CreateShareDto" }
        encryptedAlbumKey:
          type: string
          description: base64
        encryptedMediaKeys:
          type: array
          items:
            type: object
            required: [mediaId, encryptedKey, iv]
            properties:
              mediaId: { type: string, format: uuid }
              encryptedKey: { type: string, description: base64 }
              iv: { type: string, description: base64 }
        iv:
          type: string
          description: base64
        kdfParams:
          type: object
          required: [iterations, hash, salt]
          properties:
            iterations: { type: integer }
            hash: { type: string }
            salt: { type: string, description: base64 }
    CreateShareStubRequest:
      type: object
      required: [createShareDto]
      properties:
        createShareDto: { $ref: "#/components/schemas/CreateShareDto" }
    UploadShareBundleRequest:
      type: object
      required: [encryptedAlbumKey, encryptedMediaKeys, iv, kdfParams]
      properties:
        encryptedAlbumKey:
          type: string
          description: base64
        encryptedMediaKeys:
          type: array
          items:
            type: object
            required: [mediaId, encryptedKey, iv]
            properties:
              mediaId: { type: string, format: uuid }
              encryptedKey: { type: string, description: base64 }
              iv: { type: string, description: base64 }
        iv:
          type: string
          description: base64
        kdfParams:
          type: object
          required: [iterations, hash, salt]
          properties:
            iterations: { type: integer }
            hash: { type: string }
            salt: { type: string, description: base64 }
    CreateShareResponse:
      type: object
      required: [shareId, shareLink, sharePassphrase, expiresAt]
      properties:
        shareId: { type: string, format: uuid }
        shareLink: { type: string }
        sharePassphrase: { type: string }
        expiresAt: { type: string, format: date-time }
    SharedAlbumMetadata:
      type: object
      required: [shareId, album, expiresAt]
      properties:
        shareId: { type: string, format: uuid }
        album: { $ref: "#/components/schemas/Album" }
        expiresAt: { type: string, format: date-time }
        revokedAt: { type: string, format: date-time, nullable: true }
    UnlockShareRequest:
      type: object
      required: [passphrase]
      properties:
        passphrase: { type: string }
    EncryptedShareBundle:
      type: object
      required: [shareId, encryptedAlbumKey, encryptedMediaKeys, iv, kdfParams]
      properties:
        shareId: { type: string, format: uuid }
        encryptedAlbumKey: { type: string, description: base64 }
        encryptedMediaKeys:
          type: array
          items:
            type: object
            required: [mediaId, encryptedKey, iv]
            properties:
              mediaId: { type: string, format: uuid }
              encryptedKey: { type: string, description: base64 }
              iv: { type: string, description: base64 }
        iv: { type: string, description: base64 }
        kdfParams:
          type: object
          required: [iterations, hash, salt]
          properties:
            iterations: { type: integer }
            hash: { type: string }
            salt: { type: string, description: base64 }
        shareAccessToken:
          type: string
          nullable: true
          description: Short-lived token to authorize share-scoped download URLs.
        tokenExpiresAt:
          type: string
          format: date-time
          nullable: true
    SharedMediaDownloadUrl:
      type: object
      required: [url, expiresAt, method]
      properties:
        url: { type: string }
        expiresAt: { type: string, format: date-time }
        method: { type: string, example: GET }
        headers:
          type: object
          nullable: true
          additionalProperties: true
        media:
          type: object
          nullable: true
          required: [id, type, byteSize, contentType, encMeta, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            type: { type: string, enum: [PHOTO, VIDEO, DOCUMENT] }
            byteSize: { type: integer, format: int64 }
            contentType: { type: string }
            encMeta:
              type: object
              additionalProperties: true
            thumbByteSize: { type: integer, format: int64, nullable: true }
            thumbContentType: { type: string, nullable: true }
            thumbEncMeta:
              type: object
              nullable: true
              additionalProperties: true
            thumbUploadedAt: { type: string, format: date-time, nullable: true }
            originalFilename: { type: string, nullable: true }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    SharedMediaMetadataItem:
      type: object
      required: [id, type, byteSize, contentType, encMeta, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [PHOTO, VIDEO, DOCUMENT] }
        byteSize: { type: integer, format: int64 }
        contentType: { type: string }
        encMeta:
          type: object
          additionalProperties: true
        thumbByteSize: { type: integer, format: int64, nullable: true }
        thumbContentType: { type: string, nullable: true }
        thumbEncMeta:
          type: object
          nullable: true
          additionalProperties: true
        thumbUploadedAt: { type: string, format: date-time, nullable: true }
        originalFilename: { type: string, nullable: true }
        exifTakenAt: { type: string, format: date-time, nullable: true }
        takenAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SharedMediaMetadataList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SharedMediaMetadataItem" }
        nextCursor:
          type: integer
          nullable: true
          description: Cursor for the next page (index into the snapshot order).
    ShareAnalytics:
      type: object
      required: [shareId, viewCount]
      properties:
        shareId: { type: string, format: uuid }
        viewCount: { type: integer, minimum: 0 }
        lastViewedAt: { type: string, format: date-time, nullable: true }
...
    UpdateAlbumRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        coverMediaId: { type: string, format: uuid, nullable: true }
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
paths:
...
  /v1/albums/{albumId}:
    get:
      tags: [Albums]
      summary: Get album
      parameters:
        - in: path
          name: albumId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Album" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

    patch:
      tags: [Albums]
      summary: Update album
      parameters:
        - in: path
          name: albumId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAlbumRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Album" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }

    delete:
      tags: [Albums]
      summary: Delete album (soft delete)
      parameters:
        - in: path
          name: albumId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: No Content
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/share/albums/{albumId}:
    post:
      tags: [Sharing]
      summary: Create a read-only share link for an album (owner only)
      parameters:
        - in: path
          name: albumId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateShareRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [share]
                properties:
                  share: { $ref: "#/components/schemas/CreateShareResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/share/albums/{albumId}/stub:
    post:
      tags: [Sharing]
      summary: Create a share stub for an album (owner only)
      parameters:
        - in: path
          name: albumId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateShareStubRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [share]
                properties:
                  share: { $ref: "#/components/schemas/CreateShareResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/share/{shareId}:
    get:
      tags: [Sharing]
      security: []
      summary: Get shared album metadata (public)
      parameters:
        - in: path
          name: shareId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [share]
                properties:
                  share: { $ref: "#/components/schemas/SharedAlbumMetadata" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/share/{shareId}/unlock:
    post:
      tags: [Sharing]
      security: []
      summary: Unlock shared album (public; server returns encrypted bundle + short-lived share token)
      parameters:
        - in: path
          name: shareId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UnlockShareRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [bundle]
                properties:
                  bundle: { $ref: "#/components/schemas/EncryptedShareBundle" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/share/{shareId}/revoke:
    post:
      tags: [Sharing]
      summary: Revoke a share (owner only)
      parameters:
        - in: path
          name: shareId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/share/{shareId}/bundle:
    post:
      tags: [Sharing]
      summary: Upload encrypted share bundle (owner only)
      parameters:
        - in: path
          name: shareId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadShareBundleRequest" }
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/share/{shareId}/media:
    get:
      tags: [Sharing]
      security: []
      summary: List shared media metadata (public; requires X-Share-Token)
      parameters:
        - in: path
          name: shareId
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: X-Share-Token
          required: true
          schema: { type: string }
        - in: query
          name: cursor
          required: false
          schema: { type: integer }
          description: Cursor (0-based index) into the share snapshot order.
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 500 }
          description: Page size (default 200).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [media]
                properties:
                  media: { $ref: "#/components/schemas/SharedMediaMetadataList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/share/{shareId}/view:
    post:
      tags: [Sharing]
      security: []
      summary: Track a successful share view (public; requires X-Share-Token)
      parameters:
        - in: path
          name: shareId
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: X-Share-Token
          required: true
          schema: { type: string }
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/share/{shareId}/analytics:
    get:
      tags: [Sharing]
      summary: Get share analytics (owner only)
      parameters:
        - in: path
          name: shareId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [analytics]
                properties:
                  analytics: { $ref: "#/components/schemas/ShareAnalytics" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/share/albums/active:
    get:
      tags: [Sharing]
      summary: List active shares (owner only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [shares]
                properties:
                  shares:
                    type: array
                    items:
                      type: object
                      required: [id, album, expiresAt, createdAt, viewCount]
                      properties:
                        id: { type: string, format: uuid }
                        album:
                          type: object
                          required: [id, name]
                          properties:
                            id: { type: string, format: uuid }
                            name: { type: string }
                            description: { type: string, nullable: true }
                            coverMediaId: { type: string, format: uuid, nullable: true }
                        expiresAt: { type: string, format: date-time }
                        createdAt: { type: string, format: date-time }
                        viewCount: { type: integer, minimum: 0 }
                        lastViewedAt: { type: string, format: date-time, nullable: true }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/share/{shareId}/media/{mediaId}/download-url:
    post:
      tags: [Sharing]
      security: []
      summary: Get share-scoped signed download URL for a media item (public; requires X-Share-Token)
      parameters:
        - in: path
          name: shareId
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: mediaId
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: X-Share-Token
          required: true
          schema: { type: string }
        - in: query
          name: variant
          required: false
          schema:
            type: string
            enum: [original, thumb]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [downloadUrl]
                properties:
                  downloadUrl: { $ref: "#/components/schemas/SharedMediaDownloadUrl" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...